Open up JavaVirtualMachine model
The past few years have seen a variety of JVM builds get
adopted, from AdoptOpenJDK to OpenJ9 to upstream projects.
In such a scenario, the JMA built-in, explicit whitelisting
of JVMs has become a liability.

Rather than require an explicitly modeled and white-listed
JVM, this change preserves the definition of memory pools
and their mapping to configuration properties, and detects
dynamically which memory pools are available in the JVM.
The information on the supported memory pools is available
in DEBUG log level.

Also: Add a RetryRule to the e2e tests to make them less flaky,
and ignore failures in e2e tests to that all JDKs can be tested
at once.